+++ x/bridge/keeper/claim_deposit_test.go
@@ -191,8 +191,8 @@ func TestClaimDeposit(t *testing.T) {
 	aggregate := &oracletypes.Aggregate{
 		QueryId:              queryId,
 		AggregateValue:       reportValueString,
-		AggregateReportIndex: int64(0),
-		ReporterPower:        int64(68),
+		AggregateReportIndex: uint64(0),
+		ReporterPower:        uint64(68),
 	}
 	powerThreshold := uint64(67)
 	validatorTimestamp := uint64(aggregateTimestamp.UnixMilli() - 1)
@@ -201,7 +201,7 @@ func TestClaimDeposit(t *testing.T) {
 	require.NoError(t, err)
 	sdkCtx = sdkCtx.WithBlockTime(sdkCtx.BlockTime().Add(13 * time.Hour))
 	recipient, amount, _, err := k.DecodeDepositReportValue(ctx, reportValueString)
-	ok.On("GetAggregateByIndex", sdkCtx, queryId, uint64(aggregate.AggregateReportIndex)).Return(aggregate, aggregateTimestamp, err)
+	ok.On("GetAggregateByIndex", sdkCtx, queryId, (aggregate.AggregateReportIndex)).Return(aggregate, aggregateTimestamp, err)
 	bk.On("MintCoins", sdkCtx, bridgetypes.ModuleName, amount).Return(err)
 	bk.On("SendCoinsFromModuleToAccount", sdkCtx, bridgetypes.ModuleName, recipient, amount).Return(err)
 
@@ -259,15 +259,15 @@ func TestClaimDepositFlaggedAggregate(t *testing.T) {
 	aggregate := &oracletypes.Aggregate{
 		QueryId:              queryId,
 		AggregateValue:       reportValueString,
-		AggregateReportIndex: int64(0),
-		ReporterPower:        int64(90 * 1e6),
+		AggregateReportIndex: uint64(0),
+		ReporterPower:        uint64(90 * 1e6),
 		Flagged:              true,
 	}
 	msgSender := simtestutil.CreateIncrementalAccounts(2)[1]
 	sdkCtx = sdkCtx.WithBlockTime(sdkCtx.BlockTime().Add(13 * time.Hour))
 	recipient, amount, _, err := k.DecodeDepositReportValue(ctx, reportValueString)
 	totalBondedTokens := math.NewInt(100 * 1e6)
-	ok.On("GetAggregateByIndex", sdkCtx, queryId, uint64(aggregate.AggregateReportIndex)).Return(aggregate, aggregateTimestamp, err)
+	ok.On("GetAggregateByIndex", sdkCtx, queryId, aggregate.AggregateReportIndex).Return(aggregate, aggregateTimestamp, err)
 	rk.On("TotalReporterPower", sdkCtx).Return(totalBondedTokens, err)
 	bk.On("MintCoins", sdkCtx, bridgetypes.ModuleName, amount).Return(err)
 	bk.On("SendCoinsFromModuleToAccount", sdkCtx, bridgetypes.ModuleName, recipient, amount).Return(err)
@@ -308,8 +308,8 @@ func TestClaimDepositNotEnoughPower(t *testing.T) {
 	aggregate := &oracletypes.Aggregate{
 		QueryId:              queryId,
 		AggregateValue:       reportValueString,
-		AggregateReportIndex: int64(0),
-		ReporterPower:        int64(65),
+		AggregateReportIndex: uint64(0),
+		ReporterPower:        uint64(65),
 	}
 	powerThreshold := uint64(67)
 	validatorTimestamp := uint64(aggregateTimestamp.UnixMilli() - 1)
@@ -319,7 +319,7 @@ func TestClaimDepositNotEnoughPower(t *testing.T) {
 	msgSender := simtestutil.CreateIncrementalAccounts(2)[1]
 	sdkCtx = sdkCtx.WithBlockTime(sdkCtx.BlockTime().Add(13 * time.Hour))
 	recipient, amount, _, err := k.DecodeDepositReportValue(ctx, reportValueString)
-	ok.On("GetAggregateByIndex", sdkCtx, queryId, uint64(aggregate.AggregateReportIndex)).Return(aggregate, aggregateTimestamp, err)
+	ok.On("GetAggregateByIndex", sdkCtx, queryId, (aggregate.AggregateReportIndex)).Return(aggregate, aggregateTimestamp, err)
 	bk.On("MintCoins", sdkCtx, bridgetypes.ModuleName, amount).Return(err)
 	bk.On("SendCoinsFromModuleToAccount", sdkCtx, bridgetypes.ModuleName, recipient, amount).Return(err)
 
@@ -358,8 +358,8 @@ func TestClaimDepositReportTooYoung(t *testing.T) {
 	aggregate := &oracletypes.Aggregate{
 		QueryId:              queryId,
 		AggregateValue:       reportValueString,
-		AggregateReportIndex: int64(0),
-		ReporterPower:        int64(68),
+		AggregateReportIndex: uint64(0),
+		ReporterPower:        uint64(68),
 	}
 	powerThreshold := uint64(67)
 	validatorTimestamp := uint64(aggregateTimestamp.UnixMilli() - 1)
@@ -369,7 +369,7 @@ func TestClaimDepositReportTooYoung(t *testing.T) {
 	msgSender := simtestutil.CreateIncrementalAccounts(2)[1]
 	sdkCtx = sdkCtx.WithBlockTime(sdkCtx.BlockTime().Add(11 * time.Hour))
 	recipient, amount, _, err := k.DecodeDepositReportValue(ctx, reportValueString)
-	ok.On("GetAggregateByIndex", sdkCtx, queryId, uint64(aggregate.AggregateReportIndex)).Return(aggregate, aggregateTimestamp, err)
+	ok.On("GetAggregateByIndex", sdkCtx, queryId, (aggregate.AggregateReportIndex)).Return(aggregate, aggregateTimestamp, err)
 	bk.On("MintCoins", sdkCtx, bridgetypes.ModuleName, amount).Return(err)
 	bk.On("SendCoinsFromModuleToAccount", sdkCtx, bridgetypes.ModuleName, recipient, amount).Return(err)
 
@@ -408,8 +408,8 @@ func TestClaimDepositSpam(t *testing.T) {
 	aggregate := &oracletypes.Aggregate{
 		QueryId:              queryId,
 		AggregateValue:       reportValueString,
-		AggregateReportIndex: int64(0),
-		ReporterPower:        int64(68),
+		AggregateReportIndex: uint64(0),
+		ReporterPower:        uint64(68),
 	}
 	powerThreshold := uint64(67)
 	validatorTimestamp := uint64(aggregateTimestamp.UnixMilli() - 1)
@@ -419,7 +419,7 @@ func TestClaimDepositSpam(t *testing.T) {
 	msgSender := simtestutil.CreateIncrementalAccounts(2)[1]
 	sdkCtx = sdkCtx.WithBlockTime(sdkCtx.BlockTime().Add(13 * time.Hour))
 	recipient, amount, _, err := k.DecodeDepositReportValue(ctx, reportValueString)
-	ok.On("GetAggregateByIndex", sdkCtx, queryId, uint64(aggregate.AggregateReportIndex)).Return(aggregate, aggregateTimestamp, err)
+	ok.On("GetAggregateByIndex", sdkCtx, queryId, (aggregate.AggregateReportIndex)).Return(aggregate, aggregateTimestamp, err)
 	bk.On("MintCoins", sdkCtx, bridgetypes.ModuleName, amount).Return(err)
 	bk.On("SendCoinsFromModuleToAccount", sdkCtx, bridgetypes.ModuleName, recipient, amount).Return(err)
